Test: Move test function to test package: createDataStoreFactory
