+++ packages/test/local-server-tests/src/test/opsOnReconnect.spec.ts
@@ -25,12 +25,12 @@ import {
 } from "@fluidframework/local-driver/internal";
 import { SharedDirectory, type ISharedMap, SharedMap } from "@fluidframework/map/internal";
 import { FlushMode, IEnvelope } from "@fluidframework/runtime-definitions/internal";
-import { createDataStoreFactory } from "@fluidframework/runtime-utils/internal";
 import { SharedString } from "@fluidframework/sequence/internal";
 import {
 	ILocalDeltaConnectionServer,
 	LocalDeltaConnectionServer,
 } from "@fluidframework/server-local-server";
+import { createDataStoreFactory } from "@fluidframework/test-utils/internal";
 import {
 	createAndAttachContainerUsingProps,
 	ITestFluidObject,
